Fixes #3 -- Provide better feedback to user.
Previously, any time a command would run, the feedback that let the user know
the command had be started wasn't shown until after the command had been
completed. Now, the line showing the command being kicked off is printed first,
then the command is run. When the command is complete, output from the command
is shown to the user.

Going forward, we could optionally stream all stdout, and stderr as well. The
only hurdle to this is properly splitting the command prior to passing it to
child_process.spawn since it requires passing in the command, and then the
args.